- Change weights from integers to double
git-svn-id: svn://svn.sundsvall.se/oep-dev@68 fcb98548-a2c4-3c4b-99c6-acc7aeb3bb18

+++ framework/trunk/src/com/nordicpeak/flowengine/evaluators/calculatequerystateevaluator/CalculatedQueryStateEvaluationProviderModule.java
@@ -13,66 +13,64 @@ import com.nordicpeak.flowengine.interfaces.Evaluator;
 import com.nordicpeak.flowengine.interfaces.MutableEvaluatorDescriptor;
 import com.nordicpeak.flowengine.interfaces.QueryInstance;
 
-
 public abstract class CalculatedQueryStateEvaluationProviderModule<T extends CalculatedQueryStateEvaluator> extends BaseQueryStateEvaluationProviderModule<T> {
-
+	
 	@Override
 	public EvaluationResponse evaluate(QueryInstance queryInstance, User user, User poster, T evaluator, EvaluationCallback callback, boolean hasQueryValidationErrors, MutableAttributeHandler attributeHandler) {
-
+		
 		if (getTargetQueryIDs(evaluator, callback) == null || evaluator.getFromValue() == null) {
-
+			
 			return null;
 		}
 		
 		if (!supportsQueryInstance(queryInstance)) {
-
+			
 			log.warn("Query instance " + queryInstance + " is not supported for evaluator " + evaluatorDescriptor);
-
+			
 			return null;
 		}
-
-		if(hasQueryValidationErrors) {
+		
+		if (hasQueryValidationErrors) {
 			
 			return restoreDefaultQueryStates(queryInstance, evaluator, callback, attributeHandler);
 		}
 		
-		Integer calculatedValue = null;
+		Double calculatedValue = null;
 		
 		try {
-		
+			
 			calculatedValue = getCalculatedValue(queryInstance, evaluator);
-		
-		} catch(ValidationException e) {
+			
+		} catch (ValidationException e) {
 			
 			return restoreDefaultQueryStates(queryInstance, evaluator, callback, attributeHandler);
 		}
 		
-		if(calculatedValue != null && calculatedValue >= evaluator.getFromValue() && (evaluator.getToValue() == null || calculatedValue <= evaluator.getToValue())) {
+		if (calculatedValue != null && calculatedValue >= evaluator.getFromValue() && (evaluator.getToValue() == null || calculatedValue <= evaluator.getToValue())) {
 			
-			return applyEvaluatorQueryStates(queryInstance, evaluator, callback, attributeHandler);		
-		} 
+			return applyEvaluatorQueryStates(queryInstance, evaluator, callback, attributeHandler);
+		}
 		
 		return restoreDefaultQueryStates(queryInstance, evaluator, callback, attributeHandler);
-		
 	}
 	
 	@Override
 	public Evaluator createEvaluator(MutableEvaluatorDescriptor descriptor, TransactionHandler transactionHandler) throws Throwable {
 		
 		T evaluator = getEvaluatorClass().newInstance();
-
+		
 		evaluator.setEvaluatorID(descriptor.getEvaluatorID());
 		evaluator.setQueryState(QueryState.VISIBLE);
-
+		
 		getEvaluatorDAO().add(evaluator, transactionHandler, null);
-
+		
 		evaluator.init(descriptor, getFullAlias() + "/config/" + descriptor.getEvaluatorID());
-
+		
 		return evaluator;
 	}
-
-	protected abstract Integer getCalculatedValue(QueryInstance queryInstance, T evaluator) throws ValidationException;
-
+	
+	protected abstract Double getCalculatedValue(QueryInstance queryInstance, T evaluator) throws ValidationException;
+	
 	protected abstract boolean supportsQueryInstance(QueryInstance queryInstance);
 	
 }
